Move all `compile/mark-*.ts` to `compile/mark/*.ts`
