+++ src/compile/mark/text.ts
@@ -1,8 +1,8 @@
-import {Model} from './Model';
-import {X, Y, COLOR, TEXT, SIZE} from '../channel';
-import {applyMarkConfig, applyColorAndOpacity, formatMixins} from './util';
-import {extend, contains} from '../util';
-import {QUANTITATIVE, ORDINAL, TEMPORAL} from '../type';
+import {Model} from '../Model';
+import {X, Y, COLOR, TEXT, SIZE} from '../../channel';
+import {applyMarkConfig, applyColorAndOpacity, formatMixins} from '../util';
+import {extend, contains} from '../../util';
+import {QUANTITATIVE, ORDINAL, TEMPORAL} from '../../type';
 
 export namespace text {
   export function markType() {
