+++ tsconfig.json
@@ -41,13 +41,13 @@
     "src/compile/facet.ts",
     "src/compile/layout.ts",
     "src/compile/legend.ts",
-    "src/compile/mark-area.ts",
-    "src/compile/mark-bar.ts",
-    "src/compile/mark-line.ts",
-    "src/compile/mark-point.ts",
-    "src/compile/mark-text.ts",
-    "src/compile/mark-tick.ts",
-    "src/compile/mark.ts",
+    "src/compile/mark/area.ts",
+    "src/compile/mark/bar.ts",
+    "src/compile/mark/line.ts",
+    "src/compile/mark/mark.ts",
+    "src/compile/mark/point.ts",
+    "src/compile/mark/text.ts",
+    "src/compile/mark/tick.ts",
     "src/compile/Model.ts",
     "src/compile/scale.ts",
     "src/compile/stack.ts",
@@ -84,12 +84,12 @@
     "test/compile/compile.test.ts",
     "test/compile/data.test.ts",
     "test/compile/legend.test.ts",
-    "test/compile/mark-area.test.ts",
-    "test/compile/mark-bar.test.ts",
-    "test/compile/mark-line.test.ts",
-    "test/compile/mark-point.test.ts",
-    "test/compile/mark-text.test.ts",
-    "test/compile/mark-tick.test.ts",
+    "test/compile/mark/area.test.ts",
+    "test/compile/mark/bar.test.ts",
+    "test/compile/mark/line.test.ts",
+    "test/compile/mark/point.test.ts",
+    "test/compile/mark/text.test.ts",
+    "test/compile/mark/tick.test.ts",
     "test/compile/Model.test.ts",
     "test/compile/scale.test.ts",
     "test/compile/stack.test.ts",
